refactor(docker): split log stream parsing
